requirements.txt
streamlit
langchain
langchain-community
sentence-transformers
chromadb
llama-cpp-python
pypdf
python-dotenv
torch
transformers
accelerate
